+++ api/v1/class/auth/plugins/mail_sickness_approved.auth.arbeit.inc.php
@@ -1,9 +1,9 @@
 <?php
 
 namespace Arbeitszeit {
-    class MailVacationApproved extends Auth
+    class MailSicknessApproved extends Auth
     {
-        public static function mail_vacation_approved($username, $id, \PHPMailer\PHPMailer\PHPMailer $mail, $password = null)
+        public static function mail_sickness_approved($username, $id, \PHPMailer\PHPMailer\PHPMailer $mail, $password = null)
         {
             $base_url = Arbeitszeit::get_app_ini()["general"]["base_url"];
             $conn = Arbeitszeit::get_conn();
@@ -24,12 +24,12 @@ namespace Arbeitszeit {
                 ];
             }
 
-            $sql1 = "SELECT * FROM `vacation` WHERE id = '{$id}';";
+            $sql1 = "SELECT * FROM `sickness` WHERE id = '{$id}';";
             $res1 = mysqli_query($conn, $sql1);
             if ($res1 != false) {
                 $worktime_data = mysqli_fetch_assoc($res1);
             } else {
-                Exceptions::error_rep("An error occured while fetching vacation data from database for id '{$id}' | SQL-Error: " . mysqli_error($conn));
+                Exceptions::error_rep("An error occured while fetching sickness data from database for id '{$id}' | SQL-Error: " . mysqli_error($conn));
                 return [
                     "error" => [
                         "error_code" => 11,
@@ -61,7 +61,7 @@ namespace Arbeitszeit {
 </style>
                 <p>Hallo {$data["name"]},</p>
 
-                <p>dein Vorgesetzter hat deinen Urlaubsantrag <span style="color:green;">genehmigt</span>:
+                <p>dein Vorgesetzter hat deine Krankheit <span style="color:green;">genehmigt</span>:
 
                 <table>
                     <tr>
@@ -93,7 +93,7 @@ namespace Arbeitszeit {
 DATA;
             $mail->CharSet = "UTF-8";
             $mail->Encoding = "base64";
-            $mail->Subject = "Dein Urlaubsantrag wurde genehmigt!";
+            $mail->Subject = "Deine Krankheit wurde genehmigt!";
             $mail->Body = $text;
             $mail->isHTML(true);
             #try {
@@ -101,7 +101,7 @@ DATA;
             #} catch (\Exception $e){
             #  echo "Error: {$mail->ErrorInfo}";
             #}
-            Exceptions::error_rep("An email has been sent to user '$username'. Trigger: Vacation approved");
+            Exceptions::error_rep("An email has been sent to user '$username'. Trigger: Sickness approved");
             return 1;
         }
     }