+++ api/v1/class/plugins/plugins/userdetail/views/overview.php
@@ -1,0 +1,13 @@
+<?php
+require_once $_SERVER["DOCUMENT_ROOT"] . "/api/v1/inc/arbeit.inc.php";
+require_once $_SERVER["DOCUMENT_ROOT"] . "/api/v1/class/plugins/loader.plugins.arbeit.inc.php";
+require_once dirname(__DIR__, 1) . "/src/Main.php";
+
+use Userdetail\Userdetail;
+
+$main = new Userdetail;
+
+$nav = $main->compute_user_nav();
+?>
+
+<?php echo $nav ?>