+++ api/v1/class/plugins/docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v0.2
+
+* Added a permissions system, read more within the documentation
+* Added a new plugin called `userdetail`, allowing administrators to save additional data about your employees
+
 ## v0.1
 
 * Added plugin system